Library 페이징 기능 자료형 변경 및 테이블 셀 넓이 조정 등

+++ src/main/java/exam/springboot/jpa/library/DAO/LibraryDAOImpl.java
@@ -23,7 +23,7 @@ public class LibraryDAOImpl implements LibraryDAO{
 
     @Override
     public int countBoard() {
-        int allcnt = libraryRepository.countLibraryBy();
+        Long allcnt = libraryRepository.countLibraryBy();
         return (int) Math.ceil((double)allcnt/25);
     }
 }
